Made StatementBase class that both Statement and SubStatement inherit from
