+++ tincan/statement.py
@@ -16,14 +16,10 @@ import uuid
 import re
 from datetime import datetime
 
-from tincan.serializable_base import SerializableBase
+from tincan.statement_base import StatementBase
 from tincan.agent import Agent
 from tincan.group import Group
-from tincan.verb import Verb
-from tincan.attachment import Attachment
-from tincan.attachment_list import AttachmentList
 from tincan.result import Result
-from tincan.context import Context
 from tincan.substatement import SubStatement
 from tincan.statement_ref import StatementRef
 from tincan.activity import Activity
@@ -34,7 +30,7 @@ from tincan.conversions.iso8601 import make_datetime
    :synopsis: A Statement object that contains all the information for one statement that is sent to an LRS
 """
 
-class Statement(SerializableBase):
+class Statement(StatementBase):
 
     _UUID_REGEX = re.compile(
         r'^[a-f0-9]{8}-'
@@ -87,60 +83,6 @@ class Statement(SerializableBase):
         del self._id
 
     @property
-    def actor(self):
-        """Actor for Statement
-
-        :setter: Tries to convert to Agent or Group
-        :setter type: :mod:`tincan.agent` | :mod:`tincan.group`
-        :rtype: :mod:`tincan.agent` | :mod:`tincan.group`
-
-        """
-        return self._actor
-
-    @actor.setter
-    def actor(self, value):
-        if value is not None and not isinstance(value, Agent) and not isinstance(value, Group):
-            if isinstance(value, dict):
-                if 'object_type' in value or 'objectType' in value:
-                    if 'objectType' in value:
-                        value['object_type'] = value['objectType']
-                        value.pop('objectType')
-                    if value['object_type'] == 'Agent':
-                        value = Agent(value)
-                    elif value['object_type'] == 'Group':
-                        value = Group(value)
-                    else:
-                        value = Agent(value)
-                else:
-                    value = Agent(value)
-        self._actor = value
-
-    @actor.deleter
-    def actor(self):
-        del self._actor
-
-    @property
-    def verb(self):
-        """Verb for Statement
-
-        :setter: Tries to convert to Verb
-        :setter type: :mod:`tincan.verb`
-        :rtype: :mod:`tincan.verb`
-
-        """
-        return self._verb
-
-    @verb.setter
-    def verb(self, value):
-        if value is not None and not isinstance(value, Verb):
-                value = Verb(value)
-        self._verb = value
-
-    @verb.deleter
-    def verb(self):
-        del self._verb
-
-    @property
     def object(self):
         """Object for Statement
 
@@ -179,54 +121,6 @@ class Statement(SerializableBase):
     @object.deleter
     def object(self):
         del self._object
-
-    @property
-    def timestamp(self):
-        """Timestamp for Statement
-
-        :setter: Tries to convert to :class:`datetime.datetime`. If
-        no timezone is given, makes a naive `datetime.datetime`.
-
-        Strings will be parsed as ISO 8601 timestamps.
-
-        If a number is provided, it will be interpreted as a UNIX
-        timestamp, which by definition is UTC.
-
-        If a `dict` is provided, does `datetime.datetime(**value)`.
-
-        If a `tuple` or a `list` is provided, does
-        `datetime.datetime(*value)`. Uses the timezone in the tuple or
-        list if provided.
-
-        :setter type: :class:`datetime.datetime` | unicode | str | int | float | dict | tuple | list | None
-        :rtype: :class:`datetime.datetime`
-        """
-        return self._timestamp
-
-    @timestamp.setter
-    def timestamp(self, value):
-        if value is None or isinstance(value, datetime):
-            self._timestamp = value
-            return
-
-        try:
-            self._timestamp = make_datetime(value)
-        except TypeError as e:
-            e.message = (
-                "Property 'timestamp' in a 'tincan.%s' "
-                "object must be set with a "
-                "datetime.datetime, str, unicode, int, float, dict "
-                "or None.\n\n%s" %
-                (
-                    self.__class__.__name__,
-                    e.message,
-                )
-            )
-            raise e
-
-    @timestamp.deleter
-    def timestamp(self):
-        del self._timestamp
 
     @property
     def stored(self):
@@ -317,27 +211,6 @@ class Statement(SerializableBase):
         del self._result
 
     @property
-    def context(self):
-        """Context for Statement
-
-        :setter: Tries to convert to Context
-        :setter type: :mod:`tincan.context`
-        :rtype: :mod:`tincan.context`
-
-        """
-        return self._context
-
-    @context.setter
-    def context(self, value):
-        if value is not None and not isinstance(value, Context):
-                value = Context(value)
-        self._context = value
-
-    @context.deleter
-    def context(self):
-        del self._context
-
-    @property
     def version(self):
         """Version for Statement
 
@@ -360,27 +233,3 @@ class Statement(SerializableBase):
     @version.deleter
     def version(self):
         del self._version
-
-    @property
-    def attachments(self):
-        """Attachments for Statement
-
-        :setter: Tries to convert each element to Attachment
-        :setter type: :mod:`tincan.attachment_list`
-        :rtype: :mod:`tincan.attachment_list`
-
-        """
-        return self._attachments
-
-    @attachments.setter
-    def attachments(self, value):
-        if value is not None and not isinstance(value, AttachmentList):
-            try:
-                value = AttachmentList([Attachment(value)])
-            except (TypeError, AttributeError):
-                value = AttachmentList(value)
-        self._attachments = value
-
-    @attachments.deleter
-    def attachments(self):
-        del self._attachments
